Bug fix: manual_ica OSL wrapper in MNE wrapper file
